Misc tweaks and 2nd attempt at PluggableButtonMorph (longer comment follows on this)
Re: PluggableButtonMorph>>morphExtent - I read your comment and understand the issue re: icon extent and things like TaskBarMorph (which is easily fixed) but disagree with the assertion.  I can think of several use cases of more dynamic buttons where the existing approach will cause problems (a couple of examples: a taskbar with an OS X style 'genie effect', buttons displaying pictures in a scalable grid, etc.)  In the absence of better information, it seems reasonable for the initial button extent to use the icon extent and that it is up to the sender to provide the icon size they want rather than have the button prescale the source image 'on the way in' based on the current button size (which leads to ugly scaling artifacts in the case of TaskBarMorph).  Rather, it seems to make more sense to have the application prescale to the maximum size it is likely to want to use, if necessary.  Then have the button scale the destination (i.e. 'magnifiedIcon') image for drawing efficiency so that the desired button size can change over time (see TaskBarMorph and its existing x2/x4 scaling option)

In cases like this where the source image and/or the initial button size and/or the final button size may not match, why not just override the calculated extent as needed? (just as nearly everything appears to be doing under the current layout system anyway... other than window controls (and even they don't really), does anything actually use the hard-coded default values?)  In this latest changeset I've significantly pre-scaled down the TaskBarMorph icons (128@128 seemed reasonable since some people might want the x2 or x4 magnification (retina displays etc.) but even 64@64 would be a significant improvement over the current extent dimensions.  Another possiblity would be to rebuild the icons when the scale is changed and then we could scale exactly to the button extent.)  The goal of the #morphExtent changes was to try to produce an extent that will at least sometimes be what you want vs. the current state where it's almost never what you want (at least not in my experience with the current hard-coded default.)  This is why it's only calculated once and only when the extent is nil: if an extent already exists, assume whoever set it has better information.