Fix hardcoded Linux VM path in CI scripts
The path to the Linux VM in the CI scripts had the repository name
hardcoded (Cuis-Smalltalk-Dev).
This breaks the CI scripts on forks (i.e. when the repository name
changes).

The bug was fixed by using the $GITHUB_WORKSPACE variable, which
points to the location where the repository was cloned.