Deduplicate VM path in CI scripts
The path to the VM and the arguments passed to it were duplicated across
the CI scripts.

This was fixed by exporting the path to the VM and the arguments that
should be passed to it (they depend on the OS) when the VM is installed.
The runCuis.sh script was introduced to centralize the way Cuis is run.

This removed the OS checks in the installUpdates.sh and runTests.sh
scripts.